order, orderItem 완료(연동 전 목표치 완료)

+++ src/main/java/kr/co/easystock/domain/order/OrderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Order, Integer>, JpaSpecificationExecutor<Order>
+public interface OrderRepository extends JpaRepository<Order, Long>, JpaSpecificationExecutor<Order>
 {
     List<Order> findAllByUser(User user);
 }
